requirements.txt
Flask==3.0.0
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
python-dotenv==1.0.0

